feat(theme): add typography tokens
