+++ frontend/src/tokens/fontSize.ts
@@ -1,0 +1,20 @@
+export const fontSize = {
+  xs: "0.75rem",
+  sm: "0.875rem",
+  base: "1rem",
+  md: "1rem",
+  lg: "1.125rem",
+  xl: "1.25rem",
+  h6: "1.25rem",
+  h5: "1.5rem",
+  h4: "1.875rem",
+  h3: "2.25rem",
+  h2: "3rem",
+  h1: "3.75rem",
+  display1: "4.5rem",
+  display2: "6rem",
+  caption: "0.75rem",
+  button: "0.875rem",
+};
+
+export type FontSize = typeof fontSize;
